Day 18 | Completed Part 1 - Refactored Int -> Long

+++ implementation/src/main/kotlin/io/github/tomplum/aoc/aircraft/homework/Lexer.kt
@@ -14,7 +14,7 @@ class Lexer {
         while (i < data.length) {
             val token = data[i]
             when {
-                token.isDigit() -> tokens.add(Number(token.toString().toInt()))
+                token.isDigit() -> tokens.add(Number(token.toString().toLong()))
                 token == '(' -> {
                     val nested = findNestedExpressions(data.substring(i + 1), mutableListOf())
                     tokens.addAll(nested)
