+++ implementation/src/main/kotlin/io/github/tomplum/aoc/aircraft/homework/Number.kt
@@ -1,6 +1,6 @@
 package io.github.tomplum.aoc.aircraft.homework
 
-data class Number(val value: Int): Term {
+data class Number(val value: Long): Term {
     override fun solve(): Number = this
 
     override fun getLength(): Int = 1
